+++ src/main/java/com/cookCycle/repository/IngredientCategoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface IngredientCategoryRepository extends CrudRepository<IngredientCategory, Integer>{
+public interface IngredientCategoryRepository extends CrudRepository<IngredientCategory, Long>{
 
 }
