+++ src/main/java/com/cookCycle/repository/RecipeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface RecipeRepository extends CrudRepository<Recipe, Integer>{
+public interface RecipeRepository extends CrudRepository<Recipe, Long>{
 
 }
